Update faction_crime_user.rs
change progress to a float

+++ src/models/faction_crime_user.rs
@@ -20,7 +20,7 @@ pub struct FactionCrimeUser {
     pub joined_at: Option<i32>,
     /// Current planning progress on the slot.
     #[serde(rename = "progress", skip_serializing_if = "Option::is_none")]
-    pub progress: Option<i32>,
+    pub progress: Option<f32>,
 }
 
 impl FactionCrimeUser {
